api/requirements.txt
azure-functions
requests==2.32.3
azure-communication-email==1.0.0
